Link status callback conditioned with link status changed
The execution body of link status callback should be conditioned by link
status changed from activated to deactivated state and inverse. This
will avoid crash if the same callback is called twice in a row without
setting a context data in between with phLibNfc_Llcp_CheckLlcp function.
When callback is called twice in a row, the first call will free at the
end the memory allocated for context data, where a second callback will
use the same memory area. That will crash nfc process. The link status
callback is called second time to notify the upper layers that there was
a forced disconnection from remote NFC device while the transaction was
ongoing.

Change-Id: I8141919e9162c300a680b499ae612abd57a162ab
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 36010